+++ apps/client/src/components/compounds/KYCSetup/constant.ts
@@ -1,0 +1,1 @@
+export const MAX_FILE_SIZE_MB = 50
