--- apps/client/src/pages/Dashboard/kyc/KycForm.tsx
@@ -1,5 +1,0 @@
-const KycForm = () => {
-  return <div>KycForm</div>
-}
-
-export default KycForm
